requirements.txt
gspread
google-auth
google-generativeai
requests
newspaper4k
lxml_html_clean
nltk
edge-tts
gTTS
google-cloud-texttospeech
python-dotenv
apscheduler
